Move Map to atoms and tweak story
