+++ src/content-blocks/MapsBlock/index.tsx
@@ -1,5 +1,5 @@
 import BlockContainer from '@/components/atoms/BlockContainer';
-import GoogleMap from '@/components/organisms/Map';
+import GoogleMap from '@/components/atoms/Map';
 import { MapBlock as MapBlockProps } from './types';
 
 const MapsBlock = ({ title, query, location, blockProperties }: MapBlockProps): JSX.Element => {
